extracted Koamdai to other file and added its preview
